Added direct sampling to cameras
Since our lights are point like, there is a zero probability to hit it by chance. Therefore to create light paths with no scattering we need to get a bit clever. We do this by sampling only a position on the camera without direction and ask the light for a direction allowing it to handle any dirac deltas.

Also removed LenseCamera for now.