added ScenePhotonTracer and VolumePhotonTracer
these are aimed for tracing individual photons, producing countable hits (i.e. photon is either detected or not)
incorporates a wave architecture: absorption is sampled to kill photons early. After a certain number of scattering events the states are saved to memory and a new iteration is started with all threads fully converged again.