incorporated GL_EXT_ray_tracing_position_fetch
Using this extension allows to fetch the vertex position of the hit triangle directly from the TLAS skipping the corresponding manual memory fetch of vertex indices and positions. This also simplified the Geometry description list to a simple material map.

Note that since we won't fetch the normal from the geometry anymore, we now use the winding order instead.